Реорганизация snark-а
