+++ server/entities/monster_snark.h
@@ -1,0 +1,63 @@
+/***
+*
+*	Copyright (c) 1996-2002, Valve LLC. All rights reserved.
+*
+*	This product contains software technology licensed from Id
+*	Software, Inc. ("Id Technology").  Id Technology (c) 1996 Id Software, Inc.
+*	All Rights Reserved.
+*
+*   Use, distribution, and modification of this source code and/or resulting
+*   object code is restricted to non-commercial enhancements to products from
+*   Valve LLC.  All other use, distribution, or modification is prohibited
+*   without written permission from Valve LLC.
+*
+****/
+
+#pragma once
+
+#include "extdll.h"
+#include "util.h"
+#include "cbase.h"
+#include "monsters.h"
+#include "weapons.h"
+#include "nodes.h"
+#include "player.h"
+#include "soundent.h"
+#include "gamerules.h"
+#include "ggrenade.h"
+
+class CSqueakGrenade : public CGrenade
+{
+	DECLARE_CLASS(CSqueakGrenade, CGrenade);
+
+	void Spawn(void);
+	void Precache(void);
+	int  Classify(void);
+	void SuperBounceTouch(CBaseEntity *pOther);
+	void HuntThink(void);
+	int  BloodColor(void) { return BLOOD_COLOR_YELLOW; }
+	void Killed(entvars_t *pevAttacker, int iGib);
+	void GibMonster(void);
+
+	DECLARE_DATADESC();
+
+	static float m_flNextBounceSoundTime;
+
+	float m_flDie;
+	Vector m_vecTarget;
+	float m_flNextHunt;
+	float m_flNextHit;
+	Vector m_posPrev;
+	EHANDLE m_hOwner;
+	int m_iMyClass;
+};
+
+enum squeak_e
+{
+	SQUEAK_IDLE1 = 0,
+	SQUEAK_FIDGETFIT,
+	SQUEAK_FIDGETNIP,
+	SQUEAK_DOWN,
+	SQUEAK_UP,
+	SQUEAK_THROW
+};
